# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement Orange Money pour les transferts et paiements",
        "contact": {
            "email": "contact@om-pay.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://om-pay-api-2cst.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/auth/initiate-login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Initier la connexion avec numéro de téléphone",
                "operationId": "f07ec28219297e40e586ec6af71feccb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Numéro de téléphone Orange Money",
                                        "type": "string",
                                        "example": "221771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "otp": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Numéro invalide"
                    }
                }
            }
        },
        "/auth/complete-login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Compléter la connexion avec OTP",
                "operationId": "76bad780cc3103d4d6d0a59f31ddb946",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "otp"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token reçu lors de l'initiation",
                                        "type": "string"
                                    },
                                    "otp": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "phone": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "account": {
                                                    "properties": {
                                                        "account_number": {
                                                            "type": "string"
                                                        },
                                                        "balance": {
                                                            "type": "number"
                                                        },
                                                        "currency": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "OTP invalide ou expiré"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}